pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mutation-impact"
version = "0.1.0"
description = "Pipeline to model structural impact of mutations and assess harmfulness"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
	"biopython>=1.83",
	"numpy>=1.24",
	"pandas>=2.0",
	"scikit-learn>=1.3",
	"scipy>=1.10",
	"gemmi>=0.6.6",
	"matplotlib>=3.7",
	"jinja2>=3.1",
	"requests>=2.31",
	"freesasa>=2.2.1",
	"Flask>=3.0",
	"WeasyPrint>=60.0",
	"xgboost>=1.7.0",
	"seaborn>=0.12.0",
	"joblib>=1.3.0"
]

[project.scripts]
mi-validate = "mutation_impact.cli:main"
mi-run = "mutation_impact.pipeline_cli:main"
mi-web = "mutation_impact.web.app:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = [
	"mutation_impact",
	"mutation_impact.input_module",
	"mutation_impact.structure",
	"mutation_impact.features",
	"mutation_impact.classifier",
	"mutation_impact.severity",
	"mutation_impact.reporting",
	"mutation_impact.web",
	"mutation_impact.ml",
]
